+++ app/Console/Kernel.php
@@ -13,7 +13,8 @@ class Kernel extends ConsoleKernel
      * @var array
      */
     protected $commands = [
-        //
+        Commands\CompanyCreate::class,
+        Commands\UserCreate::class,
     ];
 
     /**