+++ CHANGELOG.md
@@ -8,3 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - WIP: New feature "Web Forms" will be accessible from Settings
 - New migration for web_form table, run: ```bash php artisan migrate```
 - Promote Lead to Customer working
+- We add custom console commands:
+
+Create a new company
+  ```bash
+  company:create
+  ```
+Create a new user for a company
+  ```bash
+  user:create
+  ```